refactor: narrow band-model integer types to their true widths
Band is now `u16`, bandwidth class `u8`, and the LTE feature index `u16`
throughout the raw and source sub-block model — the real domains (a plain
band stays `< NR_BAND_OFFSET`, classes are single-digit). Protobuf carries
them as `i32`, so the single decode boundary
(`RawSubBlock::from_proto_sub_block`) narrows via a new fail-closed
`narrow_field` helper: an out-of-range value errors rather than truncating,
matching the existing `ul_bw_class` presence check. Encode accessors widen
back to `i32` for the wire.

The narrowing stops at the data model on purpose: `report::combos::SubBlock`
and the shared `band_label`/`cc_class` formatters stay `i32` (they only
format a value, and the `inspect` render path is not `Result`). Documented
in DESIGN.md so it is not "unified" back into a second narrowing boundary.

Also: band labels use `CompactString` (always short); the LTE source field
is spelled `dl_feature`/`ul_feature` (no `-index`); and the `ul_feature`
omit-when-0 default (`.or(Some(0))`) is restored so `cc_to_node`'s filtered
zero stays value-identical on re-read (its loss failed the LTE scalar-
feature round-trip test).

+++ src/compiler/provision.rs
@@ -260,7 +260,7 @@ mod tests {
         }])
     }
 
-    fn nr_combo(band: i32, carriers: &[&str], skus: &[&str]) -> NrSourceCombo {
+    fn nr_combo(band: u16, carriers: &[&str], skus: &[&str]) -> NrSourceCombo {
         NrSourceCombo {
             selection: selection(carriers, skus),
             // The four corpus-verified always-`Some` header fields (all but
